+++ src/collective/compoundcriterion/queryparser.py
@@ -21,8 +21,8 @@ def _filter_is(context, row):
         for term in named_adapter.query.values():
             if not isinstance(term, dict) or \
                not 'query' in term:
-                raise ValueError("The query format is not returned by {0}"
-                                 "plone.app.querystring compliant !".format(
-                                 named_adapter))
+                raise ValueError("The query format returned by '{0}' named adapter "
+                                 "is not plone.app.querystring compliant !".format(
+                                 row.values))
         return named_adapter.query
     return {}